refactor named class
