+++ src/main/java/ru/study/servlets/AddEmployeeServlet.java
@@ -7,6 +7,7 @@ import ru.study.repository.EmployeeRepositoryCSVImpl;
 import ru.study.service.EmployeeService;
 import ru.study.service.EmployeeServiceImpl;
 
+import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletConfig;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -16,23 +17,34 @@ import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.List;
 
-@WebServlet(urlPatterns = {"/show-all-employees"}, name = "allEmployee")
-public class ShowAllEmployees extends HttpServlet {
-    private static final Logger logger = LoggerFactory.getLogger(ShowAllEmployees.class);
+@WebServlet(urlPatterns = {"/add-employees"}, name = "addEmployee")
+public class AddEmployeeServlet extends HttpServlet {
+    private static final Logger logger = LoggerFactory.getLogger(AddEmployeeServlet.class);
     private EmployeeService service;
+
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        logger.debug("Полученные запросы: {}, {}", req, resp);
+        req.getRequestDispatcher("/WEB-INF/view/addEmployee.jsp").forward(req, resp);
+    }
+
+    @Override
+    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        final Employee employee = new Employee();
+        employee.setFirstName(req.getParameter("firstName"));
+        employee.setLastName(req.getParameter("lastName"));
+        employee.setRole(req.getParameter("role"));
+        service.addEmp(employee);
+        RequestDispatcher view = req.getRequestDispatcher("/WEB-INF/view/allEmployees.jsp");
         final List<Employee> employees = service.getAllEmp();
-        logger.debug("Получен список: {}", employees);
         req.setAttribute("employees", employees);
-        req.getRequestDispatcher("/WEB-INF/view/allEmployees.jsp").forward(req, resp);
+        view.forward(req, resp);
+
     }
 
     @Override
     public void init(ServletConfig config) throws ServletException {
         super.init(config);
-        service = new EmployeeServiceImpl(new EmployeeRepositoryCSVImpl());
+        this.service = new EmployeeServiceImpl(new EmployeeRepositoryCSVImpl());
     }
 }
