+++ src/main/java/ru/study/servlets/ShowAllEmployeesServlet.java
@@ -7,7 +7,6 @@ import ru.study.repository.EmployeeRepositoryCSVImpl;
 import ru.study.service.EmployeeService;
 import ru.study.service.EmployeeServiceImpl;
 
-import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletConfig;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -17,34 +16,23 @@ import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.List;
 
-@WebServlet(urlPatterns = {"/add-employees"}, name = "addEmployee")
-public class AddEmployee extends HttpServlet {
-    private static final Logger logger = LoggerFactory.getLogger(AddEmployee.class);
+@WebServlet(urlPatterns = {"/show-all-employees"}, name = "allEmployee")
+public class ShowAllEmployeesServlet extends HttpServlet {
+    private static final Logger logger = LoggerFactory.getLogger(ShowAllEmployeesServlet.class);
     private EmployeeService service;
-
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        req.getRequestDispatcher("/WEB-INF/view/addEmployee.jsp").forward(req, resp);
-    }
-
-    @Override
-    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        final Employee employee = new Employee();
-        employee.setFirstName(req.getParameter("firstName"));
-        employee.setLastName(req.getParameter("lastName"));
-        employee.setRole(req.getParameter("role"));
-        service.addEmp(employee);
-        RequestDispatcher view = req.getRequestDispatcher("/WEB-INF/view/allEmployees.jsp");
+        logger.debug("Полученные запросы: {}, {}", req, resp);
         final List<Employee> employees = service.getAllEmp();
+        logger.debug("Получен список: {}", employees);
         req.setAttribute("employees", employees);
-        view.forward(req, resp);
-
+        req.getRequestDispatcher("/WEB-INF/view/allEmployees.jsp").forward(req, resp);
     }
 
     @Override
     public void init(ServletConfig config) throws ServletException {
         super.init(config);
-        this.service = new EmployeeServiceImpl(new EmployeeRepositoryCSVImpl());
+        service = new EmployeeServiceImpl(new EmployeeRepositoryCSVImpl());
     }
 }
